refactor(arena): split deploy_interactive.go per subcommand (#1456)
